pyproject.toml
[project]
name = "kerogen"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "django>=5.2.4",
    "djangorestframework==3.16.0",
    "pip>=25.1.1",
]
